generate fstab with UUIDs
Vagrant has experimental support for adding SATA disks. Although this
box does not support that yet (at least for the virtualbox provider,
since no AHCI storage controller with the name "SATA Controller" is
included), it is possible to add another disk. This causes a boot
failure in some circumstances because although the root is set to a UUID
in the kernel cmdline, systemd's fsck still looks for /dev/sda as stated
in /etc/fstab.