Replace packer with two custom shell scripts
Using actual VMs to build VMs is slow and error-prone (you need to use
VNC to see what is going on, and booting takes over +110 seconds as we
wait to be sure Arch Linux is ready).

build.sh can build all three images in ~135 seconds (assuming all the
packages is cached), we still need to use a VM for the actually building
in GitLab CI (as that is the only safe way it can be done at the
moment), which is a bit slower (~22 min vs ~13 min (Packer)), but that
isn't really a big issue.

In the future we can hopefully switch to Kate Containers[1] and reduce
the build time significantly.

[1] https://gitlab.archlinux.org/archlinux/infrastructure/-/issues/108