Fix boot issues caused by missing modules in the initramfs
The mkinitcpio autodetect hook removes modules not needed by the
running system from the initramfs. This make the image non-bootable
on some systems as initramfs lacks the relevant kernel modules.
Ex: Some systems need the virtio-scsi kernel module and not the
"autodetected" virtio-blk kernel module for disk access.

So for the initial install we use the fallback initramfs, and
"autodetect" should add the relevant modules to the initramfs when
the user updates the kernel.