Fix partial upgrade in the guest
A partial upgrade is often a bad idea and can cause issue like:
qemu-img: /usr/lib/libc.so.6: version `GLIBC_2.33' not found (required by qemu-img)

cow_spacesize is bumped as upgrading can use a lot of storage.

linux is ignored as we need the kernel modules for the current kernel to
be available.

Fix #136