Use DHCP for all ethernet interfaces
Change the interface match of systemd-networkd from `eth0` to `eth*` to
match all local ethernet interfaces.
This is particularly useful for Vagrant, as it requires two interfaces
(one solely used for managing the machine, another e.g. for outbound
connections) and for machines in general, as depending on their context
(e.g. plain libvirt) the user may attach more interfaces and having DHCP
available for them out-of-the-box is a quality of life improvement (as
we do not know what the user intends to do with eth0 by default).