Add libvirt-executor image for our GitLab Runner custom executor
We have provided a custom executor[1] for a few years[2] now. It uses
the "basic" image which is modified on each runner to include the
relevant packages and tweaks.

Modifying the image is cumbersome (convert qcow2 to raw, losesetup,
arch-chroot, convert raw to qcow2 etc.) and requires administrative
privileges on the runner. The arch-boxes project already has all the
needed infrastructure for doing these kinds of "modifications", so let's
just add a extra image with all the needed packages and tweaks and
remove the complexity from the infrastructure repository.

The required changes to the custom executor can be found in this
infrastructure MR[3]. This image is based on the newest version of the
"modification script" from said MR[3], just before it is removed in
favor of this image.

[1] https://docs.gitlab.com/runner/executors/custom.html
[2] https://gitlab.archlinux.org/archlinux/infrastructure/-/merge_requests/385
[3] https://gitlab.archlinux.org/archlinux/infrastructure/-/merge_requests/881