Install qemu-guest-agent in all images
QGA allows the hypervisor to perform various operations in the guest,
such as the retrieval of general resource information and user password resets.
Since most users would expect cloud web UI features that utilize this agent to
just work out of the box [1] it seems to be a good idea to install it in all
images. It adds about one MB to the image size, which is reasonable.
There is no need to enable any services, the agent will start if needed.

[1] https://docs.hetzner.com/cloud/technical-details/faq#how-are-your-images-set-up