Fix hanging socat
In socat 1.8.0.3, it appears a change was made to cause this command to hang indefinitely in case ignoreeof
is provided. As such, we're using a local timeout to kill the connection after some time has passed.

The upstream change in question is here: https://repo.or.cz/socat.git/commitdiff/fa67d7d380c2a9bbd348a86655974808ed510cfc