Remove Vagrant images
Vagrant has been removed from the repos as it is not very
maintainable[1].

Releasing new Vagrant boxes was also broken after migrating to HCP
Vagrant Registry, and upstream was not very responsive (some more
context in the previous commit[2]). We only got a single report in the 5
month period that Vagrant releasing was broken, indicating that it is
not being used much these days.

[1] https://lists.archlinux.org/archives/list/arch-dev-public@lists.archlinux.org/message/4ZG5XLMIMWQYF7ZYXNOILDZEPADUSA6H/
[2] 1d618ad ("Fix Vagrant publishing after migrating to HCP Vagrant Registry[1]")
[3] https://gitlab.archlinux.org/archlinux/arch-boxes/-/issues/171

Fix #185