Adjust Btrfs zstd compression level to 1
The default zstd compression level (3) is not optimal for NVMe SSDs,
especially when the CPU is not performant, which is often the case
for virtual machines. Adjust the compression level to 1 to strike
a balance between CPU usage and disk space savings.

The inspiration for `zstd:1` comes from Fedora Workstation.